+++ tests/provider/dwd/test_index.py
@@ -4,7 +4,10 @@ import pytest
 
 from wetterdienst import Period
 from wetterdienst.metadata.cache import CacheExpiry
-from wetterdienst.provider.dwd.observation.fileindex import _create_file_index_for_dwd_server, build_path_to_parameter
+from wetterdienst.provider.dwd.observation.fileindex import (
+    _build_url_from_dataset_and_period,
+    _create_file_index_for_dwd_server,
+)
 from wetterdienst.provider.dwd.observation.metadata import (
     DwdObservationMetadata,
 )
@@ -12,12 +15,12 @@ from wetterdienst.settings import Settings
 from wetterdienst.util.network import list_remote_files_fsspec
 
 
-def test_build_index_path():
-    path = build_path_to_parameter(
+def test__build_url_from_dataset_and_period():
+    url = _build_url_from_dataset_and_period(
         dataset=DwdObservationMetadata.daily.climate_summary,
         period=Period.HISTORICAL,
     )
-    assert path == "daily/kl/historical/"
+    assert url == "https://opendata.dwd.de/climate_environment/CDC/observations_germany/climate/daily/kl/historical/"
 
 
 @pytest.mark.remote
@@ -36,9 +39,8 @@ def test_list_files_of_climate_observations():
 @pytest.mark.remote
 def test_fileindex(default_settings):
     file_index = _create_file_index_for_dwd_server(
-        dataset=DwdObservationMetadata.daily.climate_summary,
-        period=Period.RECENT,
-        cdc_base="observations_germany/climate/",
+        url="https://opendata.dwd.de/climate_environment/CDC/observations_germany/climate/daily/kl/recent",
         settings=default_settings,
+        ttl=CacheExpiry.NO_CACHE,
     ).collect()
-    assert file_index.get_column("filename").str.contains("daily/kl/recent").all()
+    assert file_index.get_column("url").str.contains("daily/kl/recent").all()