+++ wetterdienst/provider/dwd/observation/fileindex.py
@@ -38,7 +38,7 @@ def create_file_list_for_climate_observations(
     file_index = file_index.filter(pl.col("station_id").eq(station_id))
     if date_ranges:
         file_index = file_index.filter(pl.col("date_range").is_in(date_ranges))
-    return file_index.collect().get_column("filename")
+    return file_index.collect().get_column("url")
 
 
 def create_file_index_for_climate_observations(
@@ -46,40 +46,19 @@ def create_file_index_for_climate_observations(
     period: Period,
     settings: Settings,
 ) -> pl.LazyFrame:
+    """Create a file index for a given dataset and period. The file index contains the url, filename, station id and
+    date range of the files.
     """
-    Function (cached) to create a file index of the DWD station data. The file index
-    is created for an individual set of parameters.
-    Args:
-        dataset: parameter of Parameter enumeration
-        resolution: time resolution of TimeResolution enumeration
-        period: period type of PeriodType enumeration
-    Returns:
-        file index in a pandas.DataFrame with sets of parameters and station id
-    """
-    if dataset in DWD_URBAN_DATASETS:
-        file_index = _create_file_index_for_dwd_server(
-            dataset,
-            Period.RECENT,
-            "observations_germany/climate_urban",
-            settings,
-        )
-    else:
-        file_index = _create_file_index_for_dwd_server(
-            dataset,
-            period,
-            "observations_germany/climate",
-            settings,
-        )
-
+    url = _build_url_from_dataset_and_period(dataset, period)
+    df_files = _create_file_index_for_dwd_server(url, settings, CacheExpiry.TWELVE_HOURS)
     # regarding the first filter (_), DWD has published some temporary files in the end of 2024
     # within https://opendata.dwd.de/climate_environment/CDC/observations_germany/climate/10_minutes/extreme_wind/now/
     # which are not valid files
-    file_index = file_index.filter(
+    df_files = df_files.filter(
         pl.col("filename").str.split("/").list.last().str.starts_with("_").not_(),
         pl.col("filename").str.ends_with(Extension.ZIP.value),
     )
-
-    file_index = file_index.with_columns(
+    df_files = df_files.with_columns(
         pl.col("filename")
         .str.split("/")
         .list.last()
@@ -87,13 +66,11 @@ def create_file_index_for_climate_observations(
         .str.pad_start(5, "0")
         .alias("station_id"),
     )
-
-    file_index = file_index.filter(pl.col("station_id").is_not_null() & pl.col("station_id").ne("00000"))
-
+    df_files = df_files.filter(pl.col("station_id").is_not_null() & pl.col("station_id").ne("00000"))
     if dataset.resolution.value in HIGH_RESOLUTIONS and period == Period.HISTORICAL:
         # Date range string for additional filtering of historical files
-        file_index = file_index.with_columns(pl.col("filename").str.extract(DATE_RANGE_REGEX).alias("date_range"))
-        file_index = file_index.with_columns(
+        df_files = df_files.with_columns(pl.col("filename").str.extract(DATE_RANGE_REGEX).alias("date_range"))
+        df_files = df_files.with_columns(
             pl.col("date_range")
             .str.split("_")
             .list.first()
@@ -108,8 +85,7 @@ def create_file_index_for_climate_observations(
             .map_batches(lambda dates: dates + dt.timedelta(days=1))
             .alias(Columns.END_DATE.value),
         )
-
-        file_index = file_index.with_columns(
+        df_files = df_files.with_columns(
             pl.when(pl.col(Columns.START_DATE.value) > pl.col(Columns.END_DATE.value))
             .then(pl.col(Columns.START_DATE.value).min())
             .otherwise(pl.col(Columns.START_DATE.value)),
@@ -117,59 +93,35 @@ def create_file_index_for_climate_observations(
             .then(pl.col(Columns.END_DATE.value).min())
             .otherwise(pl.col(Columns.END_DATE.value)),
         )
-
-    return file_index.sort(by=[pl.col("station_id"), pl.col("filename")])
+    return df_files.sort(by=[pl.col("station_id"), pl.col("filename")])
 
 
 def _create_file_index_for_dwd_server(
-    dataset: DatasetModel,
-    period: Period,
-    cdc_base: str,
+    url: str,
     settings: Settings,
+    ttl: CacheExpiry,
 ) -> pl.LazyFrame:
-    """
-    Function to create a file index of the DWD station data, which usually is shipped as
-    zipped/archived data. The file index is created for an individual set of parameters.
-    Args:
-        dataset: dwd dataset enumeration
-        resolution: time resolution of TimeResolution enumeration
-        period: period type of PeriodType enumeration
-        cdc_base: base path e.g. climate_observations/germany
-    Returns:
-        file index in a pandas.DataFrame with sets of parameters and station id
-    """
-    parameter_path = build_path_to_parameter(dataset, period)
-
-    url = f"https://opendata.dwd.de/climate_environment/CDC/{cdc_base}/{parameter_path}"
-
-    files_server = list_remote_files_fsspec(url, settings=settings, ttl=CacheExpiry.TWELVE_HOURS)
-
-    if not files_server:
+    urls = list_remote_files_fsspec(url, settings=settings, ttl=ttl)
+    if not urls:
         raise FileNotFoundError(f"url {url} does not have a list of files")
+    df = pl.DataFrame({"url": urls}, schema={"url": pl.String}, orient="col")
+    df = df.with_columns(
+        pl.col("url").str.split("/").list.last().alias("filename"),
+    )
+    return df.lazy()
 
-    return pl.DataFrame(files_server, schema={"filename": pl.String}, orient="col").lazy()
 
-
-def build_path_to_parameter(
+def _build_url_from_dataset_and_period(
     dataset: DatasetModel,
     period: Period,
 ) -> str:
-    """
-    Function to build a indexing file path
-    Args:
-        dataset: observation measure
-        resolution: frequency/granularity of measurement interval
-        period: recent or historical files
-
-    Returns:
-        indexing file path relative to climate observation path
-    """
-    if dataset in (
+    """Build the URL for a given dataset."""
+    if dataset in DWD_URBAN_DATASETS:
+        return f"https://opendata.dwd.de/climate_environment/CDC/observations_germany/climate_urban/{dataset.resolution.value.value}/{dataset.name_original[6:]}/{Period.RECENT.value}/"
+    elif dataset in (
         DwdObservationMetadata.hourly.solar,
         DwdObservationMetadata.daily.solar,
     ):
-        return f"{dataset.resolution.value.value}/{dataset.name_original}/"
-    elif dataset in DWD_URBAN_DATASETS:
-        return f"{dataset.resolution.value.value}/{dataset.name_original[6:]}/{period.value}/"
+        return f"https://opendata.dwd.de/climate_environment/CDC/observations_germany/climate/{dataset.resolution.value.value}/{dataset.name_original}/"
     else:
-        return f"{dataset.resolution.value.value}/{dataset.name_original}/{period.value}/"
+        return f"https://opendata.dwd.de/climate_environment/CDC/observations_germany/climate/{dataset.resolution.value.value}/{dataset.name_original}/{period.value}/"