+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ### Refactor
 - \[DWD Obs\] Make the download function more flexible using threadpool
-- \[DWD Obs\] Cleanup parser function 
+- \[DWD Obs\] Cleanup parser function
+- \[DWD Obs\] Improve fileindex and metaindex
 
 ### Fix
 - \[DWD Obs\] Reduce unnecessary file index calls during retrieval of data for stations with multiple files